refactor: schema 변경 & in절 쿼리 (#13)

+++ woorizip-adapter/web/src/main/kotlin/web/context/MemberContextHolder.kt
@@ -5,7 +5,7 @@ object MemberContextHolder {
     private val contextThread: ThreadLocal<MemberContext> = ThreadLocal.withInitial { MemberContext() }
 
     class MemberContext(
-        var memberId: Int? = null
+        var memberId: Long? = null
     )
 
     fun getContext(): MemberContext {
