+++ woorizip-core/src/main/kotlin/core/meeting/service/QueryMeetingService.kt
@@ -5,11 +5,11 @@ import core.meeting.model.Meeting
 import core.meeting.spi.QueryMeetingPort
 
 interface QueryMeetingService {
-    suspend fun getMeetingById(meetingId: Int): Meeting
+    suspend fun getMeetingById(meetingId: Long): Meeting
 }
 
 class QueryMeetingServiceImpl(private val queryMeetingPort: QueryMeetingPort) : QueryMeetingService {
-    override suspend fun getMeetingById(meetingId: Int): Meeting {
+    override suspend fun getMeetingById(meetingId: Long): Meeting {
         return queryMeetingPort.getMeetingById(meetingId) ?: throw MeetingNotFoundException(meetingId)
     }
 }
