+++ woorizip-core/src/main/kotlin/core/meeting/spi/QueryMeetingPort.kt
@@ -3,5 +3,5 @@ package core.meeting.spi
 import core.meeting.model.Meeting
 
 interface QueryMeetingPort {
-    suspend fun getMeetingById(id: Int): Meeting?
+    suspend fun getMeetingById(id: Long): Meeting?
 }
