+++ woorizip-core/src/main/kotlin/core/member/service/QueryMemberService.kt
@@ -6,12 +6,12 @@ import core.member.model.Member
 import core.member.spi.QueryMemberPort
 
 interface QueryMemberService {
-    suspend fun getMemberById(memberId: Int): Member
+    suspend fun getMemberById(memberId: Long): Member
     suspend fun getMemberByEmail(email: Email): Member
 }
 
 class QueryMemberServiceImpl(private val queryMemberPort: QueryMemberPort) : QueryMemberService {
-    override suspend fun getMemberById(memberId: Int): Member {
+    override suspend fun getMemberById(memberId: Long): Member {
         return queryMemberPort.getMemberById(memberId) ?: throw MemberNotFoundException(memberId)
     }
