+++ woorizip-core/src/main/kotlin/core/outport/TokenPort.kt
@@ -3,5 +3,5 @@ package core.outport
 import core.member.usecase.result.TokenOutput
 
 interface TokenPort {
-    suspend fun issueToken(memberId: Int): TokenOutput
+    suspend fun issueToken(memberId: Long): TokenOutput
 }
